+++ Examples/MultiPlatformTest/VolumeController.cs
@@ -6,26 +6,24 @@
     public sealed class VolumeController : BaseComponent, IUpdateableComponent {
         private AudioPlayer _audioClip;
 
-        public void Update(FrameTime frameTime) {
-            var keyboardState = Keyboard.GetState();
-
-            if (keyboardState.IsKeyDown(Keys.W)) {
+        public void Update(FrameTime frameTime, InputState inputState) {
+            if (inputState.CurrentKeyboardState.IsKeyDown(Keys.W)) {
                 this._audioClip.Volume += 0.1f;
             }
 
-            if (keyboardState.IsKeyDown(Keys.S)) {
+            if (inputState.CurrentKeyboardState.IsKeyDown(Keys.S)) {
                 this._audioClip.Volume -= 0.1f;
             }
 
-            if (keyboardState.IsKeyDown(Keys.A)) {
+            if (inputState.CurrentKeyboardState.IsKeyDown(Keys.A)) {
                 this._audioClip.Pitch += 0.1f;
             }
 
-            if (keyboardState.IsKeyDown(Keys.D)) {
+            if (inputState.CurrentKeyboardState.IsKeyDown(Keys.D)) {
                 this._audioClip.Pitch -= 0.1f;
             }
 
-            if (keyboardState.IsKeyDown(Keys.Space)) {
+            if (inputState.CurrentKeyboardState.IsKeyDown(Keys.Space)) {
                 this._audioClip.Play();
             }
         }