+++ Tests/Rendering/SpriteAnimatorTests.cs
@@ -27,13 +27,13 @@
 
             animator.Play(animation, true);
             Assert.IsNull(animator.Sprite);
-            animator.Update(frameTime);
+            animator.Update(frameTime, default);
             Assert.AreEqual(animation.Steps.ElementAt(0).Sprite.Id, animator.Sprite.Id);
 
             for (var i = 1; i < numberOfSteps; i++) {
                 gameTime.ElapsedGameTime = TimeSpan.FromSeconds(1d);
                 frameTime = new FrameTime(gameTime, 1f);
-                animator.Update(frameTime);
+                animator.Update(frameTime, default);
                 Assert.AreEqual(animation.Steps.ElementAt(i).Sprite.Id, animator.Sprite.Id);
             }
 
@@ -41,7 +41,7 @@
             frameTime = new FrameTime(gameTime, 1f);
 
             // Should loop here.
-            animator.Update(frameTime);
+            animator.Update(frameTime, default);
             Assert.AreEqual(animation.Steps.ElementAt(0).Sprite.Id, animator.Sprite.Id);
         }